build: Ignore -Wformat warnings
These warnings appeared because I use PRIu16 to print uint16_t, but on
my system it expands to "u", which is for printing unsigned int, not
unsigned short int. This is completely harmless and can only be ignored.

It sucks a bit that I won't see potential genuine warnings related to
format strings, but oh well.